+++ src/app/components/Notice.tsx
